+++ Neo Divitias/Assets/Scripts/Enemies/TurretEnemy.cs
@@ -9,7 +9,7 @@ public class TurretEnemy : DamageableObject {
     public float rotationSpeed = 1f;
     public int damageDone = 1;
     public float fireCooldown = 1f;
-    public int health = 5;
+    public float health = 5;
     public float agroDuration = 10f;
 
     public FireAtPlayer shooter;
@@ -33,7 +33,7 @@ public class TurretEnemy : DamageableObject {
         }
     }
 
-    public override void damage(int damage)
+    public override void damage(float damage)
     {
         if (!isDestructible) return;
         health -= damage;
